autofs 5.0.3 - check for exported mounts automatically mounted by kernel
If a server exports file systems that are automatically mounted by
the kernel client autofs will mistakenly over mount them when it
constructs and mounts its multi-mount triggers.

This patch makes autofs check for this case and ignores them if the
kernel mounts them while it mounts multi-mount triggers.

We don't want to fight with NFS over mounting these because it
confuses autofs and they magically go away when the owner mount is
umounted. This isn't ideal because autofs will mount these mounts
while constructing its multi-mount triggers but it is unavoidable
at the moment.