autofs-5.0.3 - fix multi source messages
There are incorrect "key not found" messages seen for master map
entries that have multiple sources (direct mounts).

For example, for the direct mount entries:

/-    auto.one
/-    auto.two

if a mount lookup is done and is not found in auto.one we see a
not found message even though it may be found in auto.two.

This patch moves the "key not found" reporting out to the higher
level lookup and reports status at the end of the lookup.