autofs-5.0.3 - clear stale flag on map read
We're not properly clearing the map stale flag after a map re-read
which causes a re-read after every lookup for master map entries
that have the "browse" option set. I removed the line that did
this at some point in the past and I must have had a reason to
do so. We'll have to wait and see what shows up after fixing
it.