autofs-5.0.3 - fix submount shutdown handling.
When using submount maps on a busy system autofs can hang.

This problem comes about because of processes walking into the
submount filesystem when it is in the process of shutting down.
While this race has been fixed for other types of mounts it
still isn't possible to to block processes from walking into
submounts that are expiring so we need to be able to recover
when this happens.

This patch improves the submount shutdown logic and allows
submounts that become busy during shutdown to recover.