autofs-5.0.3 - fix multi mount race.
When using multi-mounts it is possible for a path walk to walk into
a mount tree before it is completely setup which leads to autofs
incorrectly failing to perform mounts.

For example, for the multi-mount

mm1
     /om1       <server1>:/<path1>
     /om2       <server2>:/<path2>
     /om2/om21  <server3>:/<path3>
     /om2/om22  <server3>:/<path4>

when a path walk hits mm1/om2 <serverr2>:/<path2> is mounted on top of
mm1/om2. If a path walk comes along before the multi-mount offsets for
mm1/om2 are setup it doesn't see that mm1/om2 is pending. This happens
because the lookup gets to mm1/om2, which is within the mm1 file system,
and is covered by a mount trigger mounted at mm1/om2, and the the trigger
itself is covered by the <server3>:/<path3> mount. So the walk follows
the stack up to the mount at <server3>:/<path3>, never seeing that the
trigger mm1/om2 is currently pending.

In the example above mm1/om2 could also be a submount.

To resolve this the mount tree needs to be created under a temporary
directory and moved into place once setup in one operation.