autofs-5.0.3 - add autofs mount contol library
This patch adds a library for autofs mount point control functions.
It will use either the standard ioctl interface or, if detected, the
newer miscellaneous device file ioctl interface.

It isolates the use of autofs control functions to one place and
provides a common call interface for the daemon to use.

In addition it will check to see if the kernel supports the newer
miscellaneous device node ioctl interface and uses it if present.