autofs-5.0.3 - use miscellaneous device node, if available, for active restart
There is a problem with active restarts in autofs (that is to say
restarting autofs when there are busy mounts).

Currently autofs uses "umount -l" to clear active mounts at restart.
While using lazy umount works for most cases, anything that needs to
walk back up the mount tree to construct a path, such as getcwd(2) and
the proc file system /proc/<pid>/cwd, no longer works because the point
from which the path is constructed has been detached from the mount tree.

The actual problem with autofs is that it can't reconnect to existing
mounts. Immediately one things of just adding the ability to remount
autofs file systems would solve it, but alas, that can't work. This is
because autofs direct mounts and the implementation of "on demand mount
and expire" of nested mount trees have the file system mounted on top of
the mount trigger dentry.

To resolve this a miscellaneous device node for routing ioctl commands
to these mount points has been implemented in the autofs4 kernel module
and a library added to autofs. This provides the ability to open a file
descriptor for these over mounted autofs mount points.

This patch adds the functionaility needed to utilize these new features
and re-construct a mount tree from existing mounts and then re-connect
them. Clrearly, the maps used must correspond to the existing mounts
for this to work, so changing maps between stopping and starting
autofs again may lead to inconsistencies.