autofs-5.0.4 - fix dumb libxml2 check
The workaround for libxml2 not deleting its thread specific data
key at library unload has become broken. It was foolish of me to
think that the libxml2 folks would fix this in any reasonable time.
This patch changes the version check to 99 for the major and minor
revisions, the same as the check for the revision to ensure that
the workaround is always used and adds the setting to the build
information produced by the -V option.