autofs-5.0.5 - refactor ldap sasl bind
During the sasl authentication (and possible authentication method
selection) we establish a connection and then dispose of it and then
authenticate again. This is a little inefficient but some servers
don't like a second authentication using the same LDAP handle and
authentication fails when it should succeed. We should use the
authentication connection once we get it and not perform another
later.

Also fixed with this patch. If a server returns a set of
authentication mechanisms that all require authentication, then the
connection pointer is returned to the caller uninitialized (reported
and fix provided by Jeff Moyer).