autofs-5.0.5 - fix pidof init script usage
For some reason, following an update, pidof doesn't return the pid
of a running automount daemon when the path is used rather than just
the name, probably to do with the inode of the daemon program changing.

So we change the init script to use just the program name rather than
the path.