autofs-5.0.5 - check for path mount location in generic module
Currently we check for dependent mounts in the mount location path
for bind mounts and loopback mounts. But we can have the case where
a mount other than a loopback mount uses a local path and is not a
bind mount. In this case we need to check the local path for dependent
mounts. To do this we can check the mount location prior to spawning
the mount and if it starts with a "/" then it is a local path and
the check is needed.