autofs-5.0.5 - fix reconnect get base dn
If connect to an LDAP server but fail to find a valid query dn a
subsequent reconnect succeeds without setting the schema fields
used for constructing queries. A segfault then occurs when we try
to construct a query using the schema values that should have been
set during the query dn validation.