autofs-5.0.5 - fix get query dn failure
Recent changes to the LDAP connection logic can fail to retrieve a valid
query dn upon LDAP connection.

If LDAP is being used to store autofs maps, the LDAP schema to be used
for the maps is explicitly defined in the autofs configuration and no
master map entries exist in LDAP autofs fails to try and retrieve a
query dn, returning success instead of failure.