autofs-5.0.5 - add locality as valid ldap master map attribute fix
A recent change enabled the use of the locality ("l") attribute in map name
dns used in the master map.

When using an entry like:
/mp yp:lnxhome

the l following the ":" would match a dn patern in the tokenizer leading
to a syntax error. This has exposed a bug present for some time as, for
the map entry syntax above, this could also happen if the map name started
with another attribute, such as "cn" or "o".