autofs-5.0.5 - improve shutdown wait
When autofs manages a large number of mounts, umounting all of them
during shutdown can take a long time. The init script waits 45 seconds
for the daemon to exit, after which it reports that shutdown has failed.
But the daemon may still be busy umounting NFS mounts at this point.

The limited timeout was originally done to prevent the shutdown blocking
for a long time if a server was down. Given that the expire is now timed
we should be able to rely on /proc/mounts becoming quiesent to check if
the daemon has completed its task.