autofs-5.0.5 - don't hold lock for simple mounts
A map entry that depends on another map entry in the "same" map cannot
be allowed in all cases. In particular, multi-mount entries must not
change during mount operations so the internal map entry cache must
remain locked during the mount. This is because, during the mount, we
must consult the map and possibly update the internal cache entry to
ensure we are using the most up to date mount information.

This isn't the case for non multi-mount map entries but autofs didn't
allow for this, which is the issue this patch addresses.