autofs-5.0.5 - fix remount locking
When autofs is restarted with active mounts it is possible, due
to possible recursion when mounting multi-mount map entries, that
a lookup module will take a write lock on the map entry cache
when a read lock is alreay held.

Since, during the re-mount, we are still essentially running
single threaded we need only take care to ensure we don't take
the write lock.