autofs-5.0.5 - use null cache to mask invalid map keys
We continually get requests to have some way of avoiding mount requests
for common invalid indirect map keys. The kernel has no choice but to
pass these to the daemon and there can be many such requests at times.

This patch extends the use of null map entry cache to allow paths
corresponding to these invalid keys to be added to the master map,
checked for and ignored when a mount request is received.