autofs-5.0.5 - fix cache_delete() is to aggressive
If a mount is attempted for a non-existent map key and that key is then
added to the map and a reload performed both the negative cached key and
the newly added key were being deleted, causing mounts for the key to
fail. This is because the cache_delete() function was being to aggressive
in deleting map keys.