autofs-5.0.5 - always read file maps mount lookup map read fix
If, during a mount lookup, a file map is found to have changed and the
browse option is not being used, the file map won't be refreshed, even
though file maps should always be read into the cache.

A previous change for the "always read file maps" optimization moved
the check for the browse option into the map module read function so
checking for this in send_map_update_request() is not needed and
causes nobrowse file maps to not be read if they have changed. This
isn't quite optimal as there will be a partial read of the file map
to satisfy the lookup and a full read of the map done by the queued
update request. I don't think anything can be done about that though.