autofs-5.0.5 - fix isspace() wild card substition
If there are characters that match isspace() (such as \t, \n, etc.) in a
passed map entry key and there is no space in the key these characters
won't be properly preserved, leading to failed or incorrect mounts.

This is caused by an incorrect attempt at optimization, using a check
to see if a space is present in the passed key and only then processing
each character of the key individually, escaping any isspace() characters.