autofs-5.0.5 - fix fix direct map not updating on reread
The change to fix direct map not updating on re-read has caused
indirect maps to always be seen as stale on every mount request
lookup. This is because the map stale flag isn't being cleared
at the end of the update as is assumed by the direct map update
patch.