autofs-5.0.5 - fix map source check in file lookup
With the changes to reduce the scanning of file maps and fixes for
other issues, such as too frequent map reads the key searching has
become broken.

For an automount that has a key present in a file map and the same
key also present in another map source, if the file map entry is
removed and a lookup is performed before a re-load is issued the
map lookup fails. To fix this we need to check the map source, for
indirect maps (since plus included direct maps are handled a little
dirrefently), and continue looking if it doesn't match.