autofs-5.0.5 - fix sasl bind host name selection
A host name is needed when starting a sasl client. But the ldap
library can return a list of host names with the port appended
following a colon. With IPv6 the host name itself may have colons
so a forward search can no longer be used to find the colon
delimiting the port number and a reverse search can't be used
because it may end up including more than one host name.