autofs-5.0.5 - add nobind option
If we want to prevent the use of bind mounting for local mounts
the option "nosymlink" has to be used or we need to specify the
"port=<port number>" optionn the mount location.

To resolve this the "nobind" option is added.
The man page documentation for the "nosymlink" option is replaced
with the documentation for this option but the handling of the
"nosymlink" option has been retained for backward compatibity.