autofs-5.0.5 - dir map-type patch v2
This is the second post of "dir map-type" patch.

Changes since last post:

	- Don't use auto. as prefix for included map.
          Use .autofs suffix instead. Suggested by Steve Linn.

	- Use scandir instead of using opendir/readdir/closedir.


What is dir map-type?


`dir' map-type is for including files under a directory into master
map.
`file' map-type can be used for including a file with + notation like:

       +/etc/auto.mine

in auto.master. However, for specifying a new file to be included you
have to edit auto.master file. Editing is also needed when you want to
remove the included file. When you have to do this with your shell
script you may have to use sed or awk.

`dir' map-type permits you adding new master map entries with cp
command and removing the entries with rm command. `dir' map-type is
inspired from /etc/httpd/conf.d and /etc/modprobe.d.

`dir' map-type can be used for included files under a directory
(e.g. /etc/auto.master.d) with + notation like:

      +dir:/etc/auto.master.d

in auto.master. With this notation /etc/auto.master.d/*.autofs files
are included except a file which name is started with ".".  With the
name of the file you can control whether a file under the directory is
included or not: the file which name ends with ".autofs" is included.


Signed-off-by: Masatake YAMATO <yamato@redhat.com>