autofs-5.0.6 - fix ipv6 name lookup check
The host address must be used when the host name has multiple
addresses since we need to mount the specific host and so that
it is known what host log entries refer to.

But the check for multiple addresses can be wrong because there
is no distinction between ipv4 and ipv6 addresses. Change the
check to use the host name when mounting if neither the ipv4
or the ipv6 addresses have more than one record.