autofs-5.0.6 - fix ipv6 configure check
Since the functions clntudp6_bufcreate() and clnttcp6_create() of
libtirpc were never actually included in the library our reference
to clntudp6_bufcreate() in configure needs to be removed.

Having redone the libtirpc interface (after realizing these functions
were defined but not actually included in the library) checking for
IPv6 support can't be done now and we need to rely on the transport
independent nature of libtirpc to take care of this. So the INET6
checks have been replaced with WITH_LIBTIRPC checks instead.