autofs-5.0.6 - fix fix wait for master source mutex
The "wait for master source mutex" change wait for a busy read/write
mutex to become available instead of failing. The s390x architecture
is slower than the Intel architectures and the time allowed to wait
can be too short leading to a failure anyway.

This patch increases the maximum wait from 1 second to 5 seconds to
avoid false positive fails.