autofs-5.0.6 - fix init script usage message
Changes made to the autofs init script based on lsb compliance resulted
in a stupid change where the init script requires users to type:

/sbin/service autofs usage

to show proper usage.

Maybe it was a misunderstanding of the lsb documentation but it is still
not sensible. Change to issue the usage message when an umknown parameter
is passed but retain the return codes.