autofs-5.0.6 - reinstate singleton mount probe
The change to have the kernel process text based mount options can
introduce lengthy timeout waits when attempting a mount to a host
that is not available.

To avoid these waits autofs should probe singleton mounts if it
thinks mount.nfs will pass text options to the kernel (which of
course implies the kernel supports this).